Fixed Pop HL and Pop AF Instructions
-Both Pop HL and Pop AF instructions had a typo in the code causing the
value being input to readMemory to be left shifted by 8 instead of the
value that is returned from that function. This has been fixed for both.

+++ Source/GameBeak-Frontend/Classes/CPU.cs
@@ -4711,7 +4711,8 @@ namespace GameBeak_Frontend.Classes
         public void opcodeE1()
         {
             //Pop HL
-            Core.beakMemory.setHL(((short)(Core.beakMemory.readMemory((ushort)((Core.beakMemory.stackPointer + 1) << 8)) | (Core.beakMemory.readMemory((ushort)Core.beakMemory.stackPointer)))));
+
+            Core.beakMemory.setHL((short)((Core.beakMemory.readMemory((ushort)(Core.beakMemory.stackPointer + 1)) << 8) | (Core.beakMemory.readMemory((ushort)Core.beakMemory.stackPointer))) );
             Core.beakMemory.stackPointer += 2;
             mClock += 3;
             tClock += 12;
@@ -4842,7 +4843,7 @@ namespace GameBeak_Frontend.Classes
         public void opcodeF1()
         {
             //Pop AF
-            Core.beakMemory.setAF((Core.beakMemory.readMemory((ushort)(((Core.beakMemory.stackPointer + 1) << 8) | (Core.beakMemory.readMemory((ushort)Core.beakMemory.stackPointer))))));
+            Core.beakMemory.setAF((short)((Core.beakMemory.readMemory((ushort)(Core.beakMemory.stackPointer + 1)) << 8) | (Core.beakMemory.readMemory((ushort)Core.beakMemory.stackPointer))) );
             Core.beakMemory.stackPointer += 2;
             mClock += 3;
             tClock += 12;
